# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KTP Dashboard API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/participants:
    get:
      operationId: listEventParticipants
      tags: [participants]
      summary: List participants of an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of participants in this event
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventParticipant"
  /participants:
    get:
      operationId: listParticipants
      tags: [participants]
      summary: List all unique participants (by NIK)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Participant"
  /participants/{nik}:
    get:
      operationId: getParticipantByNik
      tags: [participants]
      summary: Get participant profile and event history
      parameters:
        - name: nik
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Participant profile with events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantProfile"
        "404":
          description: Not found
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        eventDate:
          type: string
        createdAt:
          type: string
        participantCount:
          type: integer
      required:
        - id
        - name
        - eventDate
        - createdAt
        - participantCount
...
    Participant:
      type: object
      properties:
        nik:
          type: string
        fullName:
          type: string
        address:
          type: string
          nullable: true
        birthPlace:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        religion:
          type: string
          nullable: true
        maritalStatus:
          type: string
          nullable: true
        occupation:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        eventCount:
          type: integer
        firstRegisteredAt:
          type: string
        lastRegisteredAt:
          type: string
      required:
        - nik
        - fullName
        - eventCount
        - firstRegisteredAt
        - lastRegisteredAt
    ParticipantProfile:
      type: object
      properties:
        nik:
          type: string
        fullName:
          type: string
        address:
          type: string
          nullable: true
        birthPlace:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        religion:
          type: string
          nullable: true
        maritalStatus:
          type: string
          nullable: true
        occupation:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
      required:
        - nik
        - fullName
        - events
    EventParticipant:
      type: object
      properties:
        nik:
          type: string
        fullName:
          type: string
        address:
          type: string
          nullable: true
        birthPlace:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        occupation:
          type: string
          nullable: true
        registeredAt:
          type: string
        eventCount:
          type: integer
      required:
        - nik
        - fullName
        - registeredAt
        - eventCount
...