# lib/api-zod/src/generated/types/participant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KTP Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Participant {
  nik: string;
  fullName: string;
  address?: string | null;
  birthPlace?: string | null;
  birthDate?: string | null;
  gender?: string | null;
  religion?: string | null;
  maritalStatus?: string | null;
  occupation?: string | null;
  nationality?: string | null;
  eventCount: number;
  firstRegisteredAt: string;
  lastRegisteredAt: string;
}